+++ index.html
@@ -568,8 +568,6 @@
     </dialog>
     <!-- End Modal-dialog -->
     <script type="module" src="main.js"></script>
-    <script type="text/javascript" src="challenges.js"></script>
     <script src="js/booking.js"></script>
-    <script type="module" src="challenges.js"></script>
   </body>
 </html>
